+++ packages/commands/pull.js
@@ -28,6 +28,15 @@ function pull() {
 
   const config = require(configUrl);
 
+  if (process.env.PO_APITOKEN) {
+    config.apiToken = process.env.PO_APITOKEN;
+  }
+
+  if (!config.apiToken) {
+    console.log(chalk.red(`\n API-Token not set! Most be provided in poeditor-config.json or env as "PO_APITOKEN" ~~~\n`));
+    process.exit(0);
+  }
+
   spinner = ora(`${chalk.green(`Pulling file(s) from poeditor`)}`).start();
   try {
     getTermFiles(config);